moved stackoverflow module to main package
